# Conflicts:
#	MediCon.v12.suo
#	MediCon/MediCon.csproj
#	MediCon/Models/DCIMS.edmx
#	MediCon/Models/DCIMS.edmx.diagram
#	MediCon/Models/LocalDB.edmx.diagram
#	MediCon/Models/Medicon.Context.cs
#	MediCon/Models/Medicon.Designer.cs
#	MediCon/Models/Medicon.edmx
#	MediCon/Models/OutgoingItem.cs
#	MediCon/Scripts/_references.js
#	MediCon/Web.config

+++ MediCon/Models/MenuAccess.cs
@@ -20,6 +20,6 @@ namespace MediCon.Models
         public string action { get; set; }
         public string userTypeID { get; set; }
         public string mainMenu { get; set; }
-        public Nullable<byte> orderNo { get; set; }
+        public Nullable<int> orderNo { get; set; }
     }
 }
